fix(popover): 'destroyOutsideClickHandler is not a function' (#95)

+++ src/LumexUI/wwwroot/js/components/popover.bundle.js
@@ -1475,7 +1475,10 @@ async function initialize(id, options) {
 }
 
 function destroy() {
-    destroyOutsideClickHandler();
+    if (destroyOutsideClickHandler) {
+        destroyOutsideClickHandler();
+        destroyOutsideClickHandler = null;
+    }
 }
 
 const popover = {
